+++ esp-radio/src/ble/btdm.rs
@@ -1,6 +1,7 @@
 use alloc::boxed::Box;
 use core::ptr::{addr_of, addr_of_mut};
 
+use esp_phy::{PhyController, PhyInitGuard};
 use esp_sync::RawMutex;
 use esp_wifi_sys::c_types::*;
 use portable_atomic::{AtomicBool, Ordering};
@@ -53,6 +54,9 @@ unsafe extern "C" {
     fn API_vhci_host_check_send_available() -> bool;
     fn API_vhci_host_send_packet(data: *const u8, len: u16);
     fn API_vhci_host_register_callback(vhci_host_callbac: *const VhciHostCallbacks) -> i32;
+
+    #[cfg(not(esp32))]
+    fn coex_pti_v2();
 }
 
 static VHCI_HOST_CALLBACK: VhciHostCallbacks = VhciHostCallbacks {
@@ -299,7 +303,8 @@ unsafe extern "C" fn custom_queue_create(
     todo!();
 }
 
-pub(crate) fn ble_init() {
+pub(crate) fn ble_init() -> PhyInitGuard<'static> {
+    let phy_init_guard;
     unsafe {
         (*addr_of_mut!(HCI_OUT_COLLECTOR)).write(HciOutCollector::new());
         // turn on logging
@@ -352,16 +357,19 @@ pub(crate) fn ble_init() {
         #[cfg(coex)]
         crate::binary::include::coex_enable();
 
-        crate::common_adapter::phy_enable();
+        phy_init_guard = esp_hal::peripherals::BT::steal().enable_phy();
 
-        #[cfg(esp32)]
-        {
-            unsafe extern "C" {
-                fn btdm_rf_bb_init_phase2();
+        cfg_if::cfg_if! {
+            if #[cfg(esp32)] {
+                unsafe extern "C" {
+                    fn btdm_rf_bb_init_phase2();
+                }
+
+                btdm_rf_bb_init_phase2();
+                coex_bt_high_prio();
+            } else {
+                coex_pti_v2();
             }
-
-            btdm_rf_bb_init_phase2();
-            coex_bt_high_prio();
         }
 
         #[cfg(coex)]
@@ -374,6 +382,7 @@ pub(crate) fn ble_init() {
 
     // At some point the "High-speed ADC" entropy source became available.
     unsafe { esp_hal::rng::TrngSource::increase_entropy_source_counter() };
+    phy_init_guard
 }
 
 pub(crate) fn ble_deinit() {
@@ -387,8 +396,8 @@ pub(crate) fn ble_deinit() {
 
     unsafe {
         btdm_controller_deinit();
-        crate::common_adapter::phy_disable();
     }
+    // Disabling the PHY happens automatically, when the BLEController gets dropped.
 }
 /// Sends HCI data to the BLE controller.
 #[instability::unstable]