+++ esp-radio/src/wifi/mod.rs
@@ -1370,7 +1370,20 @@ pub fn sta_mac() -> [u8; 6] {
     mac
 }
 
-pub(crate) fn wifi_init() -> Result<(), WifiError> {
+#[cfg(esp32)]
+fn set_mac_time_update_cb(wifi: crate::hal::peripherals::WIFI<'_>) {
+    use esp_phy::MacTimeExt;
+    use esp_wifi_sys::include::esp_wifi_internal_update_mac_time;
+    unsafe {
+        wifi.set_mac_time_update_cb(|duration| {
+            esp_wifi_internal_update_mac_time(duration.as_micros() as u32);
+        });
+    }
+}
+
+pub(crate) fn wifi_init(_wifi: crate::hal::peripherals::WIFI<'_>) -> Result<(), WifiError> {
+    #[cfg(esp32)]
+    set_mac_time_update_cb(_wifi);
     unsafe {
         #[cfg(coex)]
         esp_wifi_result!(coex_init())?;
@@ -1392,12 +1405,6 @@ pub(crate) fn wifi_init() -> Result<(), WifiError> {
             esp_interface_t_ESP_IF_WIFI_AP,
             Some(recv_cb_ap)
         ))?;
-
-        #[cfg(any(esp32, esp32s3))]
-        {
-            static mut NVS_STRUCT: [u32; 12] = [0; 12];
-            chip_specific::__ESP_RADIO_G_MISC_NVS = addr_of!(NVS_STRUCT) as u32;
-        }
 
         Ok(())
     }
@@ -2951,7 +2958,7 @@ impl WifiConfig {
 /// currently in use.
 pub fn new<'d>(
     _inited: &'d Controller<'d>,
-    _device: crate::hal::peripherals::WIFI<'d>,
+    device: crate::hal::peripherals::WIFI<'d>,
     config: WifiConfig,
 ) -> Result<(WifiController<'d>, Interfaces<'d>), WifiError> {
     if crate::is_interrupts_disabled() {
@@ -2998,7 +3005,7 @@ pub fn new<'d>(
         TX_QUEUE_SIZE.store(config.tx_queue_size, Ordering::Relaxed);
     };
 
-    crate::wifi::wifi_init()?;
+    crate::wifi::wifi_init(device)?;
 
     unsafe {
         let country = config.country_code.into_blob();